+++ index.html
@@ -182,8 +182,7 @@
       </div>
     </section>
   </main>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 
   <footer class="page-footer  container">
     <div class="page-footer__wrapper">
